Load and prepare DB before test
The bin/rake db:migrate task runs the migrations and then dumps the
structure of the database to a file called db/schema.rb. This structure
allows you to restore your database using the bin/rake db:schema:load
task if you wish, which is better than running all the migrations on a
large project again!

Source:
http://www.42.mach7x.com/2013/10/03/dbmigrate-or-dbschemaload-for-rails-
project-with-many-migations/