Migration check differences in Rails 4.0 & 4.1+
In Rails 4.0.x add `ActiveRecord::Migration.check_pending!` to the top
of the rails_helper file after Rails has been required. This will raise
an exception if there are any pending schema changes. Users will still
be required to manually keep the development and test environments in
sync.

With Rails 4.1+ there was an exciting new feature. Users no longer need
to keep the development and test environments in sync. To take
advantage of this add `ActiveRecord::Migration.maintain_test_schema!`
to the top of the rails_helper file after Rails has been required.