Update Travis-CI migration commands
The engine contains migrations for the annotator_store_annotations and
annotator_store_ranges table which need to be created in the
application's database so that the engine's models can query them
correctly. We now copy these migrations into the application. Only copy
over migrations that haven't been copied over already